==> All_of_Us_Researcher_Workbench/03_AoU_pull_drugs_and_clean_files.ipynb <==
import pandas
import os

# This query represents dataset "all_drugs_to_replicate_NOV_05" for domain "drug" and was generated for All of Us Registered Tier Dataset v8
dataset_30055620_drug_sql = """
    SELECT
        d_exposure.person_id,
        d_standard_concept.concept_name as standard_concept_name,
        d_exposure.drug_exposure_start_datetime 
    FROM
        ( SELECT
            * 
        FROM
            `""" + os.environ["WORKSPACE_CDR"] + """.drug_exposure` d_exposure 
        WHERE
            (
                drug_concept_id IN (SELECT
                    DISTINCT ca.descendant_id 
                FROM
                    `""" + os.environ["WORKSPACE_CDR"] + """.cb_criteria_ancestor` ca 
                JOIN
                    (SELECT
                        DISTINCT c.concept_id       
                    FROM
                        `""" + os.environ["WORKSPACE_CDR"] + """.cb_criteria` c       
                    JOIN
                        (SELECT
                            CAST(cr.id as string) AS id             
                        FROM
                            `""" + os.environ["WORKSPACE_CDR"] + """.cb_criteria` cr             
                        WHERE
                            concept_id IN (1101554, 1103314, 1106776, 1107830, 1110410, 1112807, 1112921, 1115008, 1115572, 1118084, 1124300, 1124957, 1125315, 1136422, 1137529, 1140643, 1149196, 1149380, 1150345, 1153013, 1153428, 1154161, 1154332, 1167322, 1177480, 1178663, 1189596, 1189754, 1195492, 1197736, 1201620, 1236744, 1303425, 1305058, 1307046, 1307863, 1308216, 1308842, 1309944, 1310149, 1311276, 1314002, 1316262, 1316354, 1318853, 1322184, 1326303, 1328165, 1331270, 1332418, 1334456, 1336926, 1338005, 1341927, 1347384, 1348265, 1351557, 1353766, 1353776, 1363053, 1367500, 1370109, 1373225, 1398937, 1436678, 1501700, 1503297, 1510813, 1511349, 1516800, 1525215, 1526475, 1539403, 1545958, 1547504, 1549686, 1550557, 1551803, 1551860, 1555120, 1557272, 1580747, 1597756, 1703603, 1703653, 1703687, 1705674, 1707164, 1713332, 1717240, 1728416, 1729720, 1734104, 1738521, 1746940, 1750500, 1754994, 1759842, 1760616, 1768849, 1777087, 1786621, 1786842, 1797513, 1836948, 19020124,
 19027958, 19037833, 19045317, 19080432, 19086491, 19087090, 19124477, 19127669, 21600713, 21602056, 21602797, 21602798, 21602818, 21602819, 21602840, 21602852, 21602861, 21602929, 21602930, 21602960, 21604555, 40241331, 42873636, 43013024, 701322, 703547, 704943, 705103, 710062, 711584, 713823, 715233, 715259, 715710, 715939, 715997, 720810, 721724, 722031, 723013, 725131, 733301, 733523, 734354, 735979, 739138, 740275, 740560, 740910, 741530, 743670, 744740, 745466, 750982, 751246, 752061, 755695, 757352, 757627, 766209, 766529, 766814, 767410, 777221, 778268, 782211, 785788, 789578, 791967, 794852, 797399, 797617, 798834, 798874, 836715, 904453, 906780, 908464, 909358, 911735, 913782, 914335, 916005, 918906, 920293, 922570, 923081, 923645, 924566, 924724, 924939, 925952, 929887, 930021, 932745, 939259, 941258, 948078, 948487, 948515, 952045, 956874, 960820, 961047, 964339, 968426, 970250, 978555, 985247, 987245, 989482, 989878, 990678, 991382, 991825, 991876, 996416,
 997276)             
                            AND full_text LIKE '%_rank1]%'       ) a 
                            ON (c.path LIKE CONCAT('%.', a.id, '.%') 
                            OR c.path LIKE CONCAT('%.', a.id) 
                            OR c.path LIKE CONCAT(a.id, '.%') 
                            OR c.path = a.id) 
                    WHERE
                        is_standard = 1 
                        AND is_selectable = 1) b 
                        ON (ca.ancestor_id = b.concept_id)))  
                    AND (d_exposure.PERSON_ID IN (SELECT
                        distinct person_id  
                FROM
                    `""" + os.environ["WORKSPACE_CDR"] + """.cb_search_person` cb_search_person  
                WHERE
                    cb_search_person.person_id IN (SELECT
                        criteria.person_id 
                    FROM
                        (SELECT
                            DISTINCT person_id, entry_date, concept_id 
                        FROM
                            `""" + os.environ["WORKSPACE_CDR"] + """.cb_search_all_events` 
                        WHERE
                            (concept_id IN(SELECT
                                DISTINCT c.concept_id 
                            FROM
                                `""" + os.environ["WORKSPACE_CDR"] + """.cb_criteria` c 
                            JOIN
                                (SELECT
                                    CAST(cr.id as string) AS id       
                                FROM
                                    `""" + os.environ["WORKSPACE_CDR"] + """.cb_criteria` cr       
                                WHERE
                                    concept_id IN (378419)       
                                    AND full_text LIKE '%_rank1]%'      ) a 
                                    ON (c.path LIKE CONCAT('%.', a.id, '.%') 
                                    OR c.path LIKE CONCAT('%.', a.id) 
                                    OR c.path LIKE CONCAT(a.id, '.%') 
                                    OR c.path = a.id) 
                            WHERE
                                is_standard = 1 
                                AND is_selectable = 1) 
                            AND is_standard = 1 )) criteria ) 
                    AND cb_search_person.person_id IN (SELECT
                        person_id 
                    FROM
                        `""" + os.environ["WORKSPACE_CDR"] + """.person` p 
                    WHERE
                        race_concept_id IN (8527) ) 
                    AND cb_search_person.person_id NOT IN (SELECT
                        criteria.person_id 
                    FROM
                        (SELECT
                            DISTINCT person_id, entry_date, concept_id 
                        FROM
                            `""" + os.environ["WORKSPACE_CDR"] + """.cb_search_all_events` 
                        WHERE
                            (concept_id IN(SELECT
                                DISTINCT c.concept_id 
                            FROM
                                `""" + os.environ["WORKSPACE_CDR"] + """.cb_criteria` c 
                            JOIN
                                (SELECT
                                    CAST(cr.id as string) AS id       
                                FROM
                                    `""" + os.environ["WORKSPACE_CDR"] + """.cb_criteria` cr       
                                WHERE
                                    concept_id IN (1568295)       
                                    AND full_text LIKE '%_rank1]%'      ) a 
                                    ON (c.path LIKE CONCAT('%.', a.id, '.%') 
                                    OR c.path LIKE CONCAT('%.', a.id) 
                                    OR c.path LIKE CONCAT(a.id, '.%') 
                                    OR c.path = a.id) 
                            WHERE
                                is_standard = 0 
                                AND is_selectable = 1) 
                            AND is_standard = 0 )) criteria 
                    UNION
                    DISTINCT SELECT
                        criteria.person_id 
                    FROM
                        (SELECT
                            DISTINCT person_id, entry_date, concept_id 
                        FROM
                            `""" + os.environ["WORKSPACE_CDR"] + """.cb_search_all_events` 
                        WHERE
                            (concept_id IN (373182) 
                            AND is_standard = 1 )) criteria 
                    UNION
                    DISTINCT SELECT
                        criteria.person_id 
                    FROM
                        (SELECT
                            DISTINCT person_id, entry_date, concept_id 
                        FROM
                            `""" + os.environ["WORKSPACE_CDR"] + """.cb_search_all_events` 
                        WHERE
                            (concept_id IN(SELECT
                                DISTINCT c.concept_id 
                            FROM
                                `""" + os.environ["WORKSPACE_CDR"] + """.cb_criteria` c 
                            JOIN
                                (SELECT
                                    CAST(cr.id as string) AS id       
                                FROM
                                    `""" + os.environ["WORKSPACE_CDR"] + """.cb_criteria` cr       
                                WHERE
                                    concept_id IN (381270)       
                                    AND full_text LIKE '%_rank1]%'      ) a 
                                    ON (c.path LIKE CONCAT('%.', a.id, '.%') 
                                    OR c.path LIKE CONCAT('%.', a.id) 
                                    OR c.path LIKE CONCAT(a.id, '.%') 
                                    OR c.path = a.id) 
                            WHERE
                                is_standard = 1 
                                AND is_selectable = 1) 
                            AND is_standard = 1 )) criteria 
                    UNION
                    DISTINCT SELECT
                        criteria.person_id 
                    FROM
                        (SELECT
                            DISTINCT person_id, entry_date, concept_id 
                        FROM
                            `""" + os.environ["WORKSPACE_CDR"] + """.cb_search_all_events` 
                        WHERE
                            (concept_id IN(SELECT
                                DISTINCT c.concept_id 
                            FROM
                                `""" + os.environ["WORKSPACE_CDR"] + """.cb_criteria` c 
                            JOIN
                                (SELECT
                                    CAST(cr.id as string) AS id       
                                FROM
                                    `""" + os.environ["WORKSPACE_CDR"] + """.cb_criteria` cr       
                                WHERE
                                    concept_id IN (4140090)       
                                    AND full_text LIKE '%_rank1]%'      ) a 
                                    ON (c.path LIKE CONCAT('%.', a.id, '.%') 
                                    OR c.path LIKE CONCAT('%.', a.id) 
                                    OR c.path LIKE CONCAT(a.id, '.%') 
                                    OR c.path = a.id) 
                            WHERE
                                is_standard = 1 
                                AND is_selectable = 1) 
                            AND is_standard = 1 )) criteria 
                    UNION
                    DISTINCT SELECT
                        criteria.person_id 
                    FROM
                        (SELECT
                            DISTINCT person_id, entry_date, concept_id 
                        FROM
                            `""" + os.environ["WORKSPACE_CDR"] + """.cb_search_all_events` 
                        WHERE
                            (concept_id IN (40391011) 
                            AND is_standard = 1 )) criteria 
                    UNION
                    DISTINCT SELECT
                        criteria.person_id 
                    FROM
                        (SELECT
                            DISTINCT person_id, entry_date, concept_id 
                        FROM
                            `""" + os.environ["WORKSPACE_CDR"] + """.cb_search_all_events` 
                        WHERE
                            (concept_id IN (4044053) 
                            AND is_standard = 1 )) criteria )  
                UNION
                DISTINCT SELECT
                    distinct person_id  
                FROM
                    `""" + os.environ["WORKSPACE_CDR"] + """.cb_search_person` cb_search_person  
                WHERE
                    cb_search_person.person_id IN (SELECT
                        criteria.person_id 
                    FROM
                        (SELECT
                            DISTINCT person_id, entry_date, concept_id 
                        FROM
                            `""" + os.environ["WORKSPACE_CDR"] + """.cb_search_all_events` 
                        WHERE
                            (concept_id IN(SELECT
                                DISTINCT c.concept_id 
                            FROM
                                `""" + os.environ["WORKSPACE_CDR"] + """.cb_criteria` c 
                            JOIN
                                (SELECT
                                    CAST(cr.id as string) AS id       
                                FROM
                                    `""" + os.environ["WORKSPACE_CDR"] + """.cb_criteria` cr       
                                WHERE
                                    concept_id IN (381270)       
                                    AND full_text LIKE '%_rank1]%'      ) a 
                                    ON (c.path LIKE CONCAT('%.', a.id, '.%') 
                                    OR c.path LIKE CONCAT('%.', a.id) 
                                    OR c.path LIKE CONCAT(a.id, '.%') 
                                    OR c.path = a.id) 
                            WHERE
                                is_standard = 1 
                                AND is_selectable = 1) 
                            AND is_standard = 1 )) criteria ) 
                    AND cb_search_person.person_id IN (SELECT
                        person_id 
                    FROM
                        `""" + os.environ["WORKSPACE_CDR"] + """.person` p 
                    WHERE
                        race_concept_id IN (8527) ) 
                    AND cb_search_person.person_id NOT IN (SELECT
                        criteria.person_id 
                    FROM
                        (SELECT
                            DISTINCT person_id, entry_date, concept_id 
                        FROM
                            `""" + os.environ["WORKSPACE_CDR"] + """.cb_search_all_events` 
                        WHERE
                            (concept_id IN (4043378) 
                            AND is_standard = 1 )) criteria 
                    UNION
                    DISTINCT SELECT
                        criteria.person_id 
                    FROM
                        (SELECT
                            DISTINCT person_id, entry_date, concept_id 
                        FROM
                            `""" + os.environ["WORKSPACE_CDR"] + """.cb_search_all_events` 
                        WHERE
                            (concept_id IN(SELECT
                                DISTINCT c.concept_id 
                            FROM
                                `""" + os.environ["WORKSPACE_CDR"] + """.cb_criteria` c 
                            JOIN
                                (SELECT
                                    CAST(cr.id as string) AS id       
                                FROM
                                    `""" + os.environ["WORKSPACE_CDR"] + """.cb_criteria` cr       
                                WHERE
                                    concept_id IN (443605)       
                                    AND full_text LIKE '%_rank1]%'      ) a 
                                    ON (c.path LIKE CONCAT('%.', a.id, '.%') 
                                    OR c.path LIKE CONCAT('%.', a.id) 
                                    OR c.path LIKE CONCAT(a.id, '.%') 
                                    OR c.path = a.id) 
                            WHERE
                                is_standard = 1 
                                AND is_selectable = 1) 
                            AND is_standard = 1 )) criteria 
                    UNION
                    DISTINCT SELECT
                        criteria.person_id 
                    FROM
                        (SELECT
                            DISTINCT person_id, entry_date, concept_id 
                        FROM
                            `""" + os.environ["WORKSPACE_CDR"] + """.cb_search_all_events` 
                        WHERE
                            (concept_id IN(SELECT
                                DISTINCT c.concept_id 
                            FROM
                                `""" + os.environ["WORKSPACE_CDR"] + """.cb_criteria` c 
                            JOIN
                                (SELECT
                                    CAST(cr.id as string) AS id       
                                FROM
                                    `""" + os.environ["WORKSPACE_CDR"] + """.cb_criteria` cr       
                                WHERE
                                    concept_id IN (378419)       
                                    AND full_text LIKE '%_rank1]%'      ) a 
                                    ON (c.path LIKE CONCAT('%.', a.id, '.%') 
                                    OR c.path LIKE CONCAT('%.', a.id) 
                                    OR c.path LIKE CONCAT(a.id, '.%') 
                                    OR c.path = a.id) 
                            WHERE
                                is_standard = 1 
                                AND is_selectable = 1) 
                            AND is_standard = 1 )) criteria 
                    UNION
                    DISTINCT SELECT
                        criteria.person_id 
                    FROM
                        (SELECT
                            DISTINCT person_id, entry_date, concept_id 
                        FROM
                            `""" + os.environ["WORKSPACE_CDR"] + """.cb_search_all_events` 
                        WHERE
                            (concept_id IN (373182) 
                            AND is_standard = 1 )) criteria 
                    UNION
                    DISTINCT SELECT
                        criteria.person_id 
                    FROM
                        (SELECT
                            DISTINCT person_id, entry_date, concept_id 
                        FROM
                            `""" + os.environ["WORKSPACE_CDR"] + """.cb_search_all_events` 
                        WHERE
                            (concept_id IN (40391011) 
                            AND is_standard = 1 )) criteria 
                    UNION
                    DISTINCT SELECT
                        criteria.person_id 
                    FROM
                        (SELECT
                            DISTINCT person_id, entry_date, concept_id 
                        FROM
                            `""" + os.environ["WORKSPACE_CDR"] + """.cb_search_all_events` 
                        WHERE
                            (concept_id IN (4044053) 
                            AND is_standard = 1 )) criteria )  
                UNION
                DISTINCT SELECT
                    distinct person_id  
                FROM
                    `""" + os.environ["WORKSPACE_CDR"] + """.cb_search_person` cb_search_person  
                WHERE
                    cb_search_person.person_id IN (SELECT
                        criteria.person_id 
                    FROM
                        (SELECT
                            DISTINCT person_id, entry_date, concept_id 
                        FROM
                            `""" + os.environ["WORKSPACE_CDR"] + """.cb_search_all_events` 
                        WHERE
                            (concept_id IN(SELECT
                                DISTINCT c.concept_id 
                            FROM
                                `""" + os.environ["WORKSPACE_CDR"] + """.cb_criteria` c 
                            JOIN
                                (SELECT
                                    CAST(cr.id as string) AS id       
                                FROM
                                    `""" + os.environ["WORKSPACE_CDR"] + """.cb_criteria` cr       
                                WHERE
                                    concept_id IN (4182210)       
                                    AND full_text LIKE '%_rank1]%'      ) a 
                                    ON (c.path LIKE CONCAT('%.', a.id, '.%') 
                                    OR c.path LIKE CONCAT('%.', a.id) 
                                    OR c.path LIKE CONCAT(a.id, '.%') 
                                    OR c.path = a.id) 
                            WHERE
                                is_standard = 1 
                                AND is_selectable = 1) 
                            AND is_standard = 1 )) criteria ) 
                    AND cb_search_person.person_id IN (SELECT
                        person_id 
                    FROM
                        `""" + os.environ["WORKSPACE_CDR"] + """.person` p 
                    WHERE
                        race_concept_id IN (8527) )  
                UNION
                DISTINCT SELECT
                    distinct person_id  
                FROM
                    `""" + os.environ["WORKSPACE_CDR"] + """.cb_search_person` cb_search_person  
                WHERE
                    cb_search_person.person_id IN (SELECT
                        person_id 
                    FROM
                        `""" + os.environ["WORKSPACE_CDR"] + """.cb_search_person` p 
                    WHERE
                        DATE_DIFF(CURRENT_DATE, dob, YEAR) - IF(EXTRACT(MONTH FROM dob)*100 + EXTRACT(DAY FROM dob) > EXTRACT(MONTH FROM CURRENT_DATE)*100 + EXTRACT(DAY FROM CURRENT_DATE), 1, 0) BETWEEN 60 AND 91 
                        AND NOT EXISTS (      SELECT
                            'x'      
                        FROM
                            `""" + os.environ["WORKSPACE_CDR"] + """.death` d      
                        WHERE
                            d.person_id = p.person_id ) ) 
                    AND cb_search_person.person_id IN (SELECT
                        person_id 
                    FROM
                        `""" + os.environ["WORKSPACE_CDR"] + """.person` p 
                    WHERE
                        race_concept_id IN (8527) ) 
                    AND cb_search_person.person_id NOT IN (SELECT
                        criteria.person_id 
                    FROM
                        (SELECT
                            DISTINCT person_id, entry_date, concept_id 
                        FROM
                            `""" + os.environ["WORKSPACE_CDR"] + """.cb_search_all_events` 
                        WHERE
                            (concept_id IN(SELECT
                                DISTINCT c.concept_id 
                            FROM
                                `""" + os.environ["WORKSPACE_CDR"] + """.cb_criteria` c 
                            JOIN
                                (SELECT
                                    CAST(cr.id as string) AS id       
                                FROM
                                    `""" + os.environ["WORKSPACE_CDR"] + """.cb_criteria` cr       
                                WHERE
                                    concept_id IN (44831122, 1568428, 45547675, 44826540, 1568088, 1568429, 1568284, 1568360, 1568087, 1568442, 35207459, 35207114, 1568372, 1568370, 35207329, 44819535, 1568441, 45547690, 1568408, 44819583, 1568297, 44824105, 1568279, 1568299, 1568423, 1568397, 1568298, 35207440, 1568282, 35207445, 1568406, 1568294, 1568415, 1568361, 1568394, 920130, 44821814, 1568293, 1568395, 35207501, 45566776, 45890911, 1568407, 1568400, 1568404, 1568393, 1568436, 1568286, 45552458, 44828400, 35207314, 35207429, 1568374, 1568280, 35207365, 35207460, 44830700, 1568435, 35207511, 44820755, 44826489, 1568373, 1568287, 1568412, 44830618, 44833435, 1568289, 44826537, 1568382, 1568402)       
                                    AND full_text LIKE '%_rank1]%'      ) a 
                                    ON (c.path LIKE CONCAT('%.', a.id, '.%') 
                                    OR c.path LIKE CONCAT('%.', a.id) 
                                    OR c.path LIKE CONCAT(a.id, '.%') 
                                    OR c.path = a.id) 
                            WHERE
                                is_standard = 0 
                                AND is_selectable = 1) 
                            AND is_standard = 0 )) criteria ) )
            )
        ) d_exposure 
    LEFT JOIN
        `""" + os.environ["WORKSPACE_CDR"] + """.concept` d_standard_concept 
            ON d_exposure.drug_concept_id = d_standard_concept.concept_id"""

dataset_30055620_drug_df = pandas.read_gbq(
    dataset_30055620_drug_sql,
    dialect="standard",
    use_bqstorage_api=("BIGQUERY_STORAGE_API_ENABLED" in os.environ),
    progress_bar_type="tqdm_notebook")

dataset_30055620_drug_df.head(5)

#2 million more meds with v8
meds = dataset_30055620_drug_df
meds.shape

# Select the columns we need
meds = meds[['person_id', 'standard_concept_name', 'drug_exposure_start_datetime']]
meds

# Split date-time to get date
meds = meds.rename(columns = {'person_id':'ID', 'standard_concept_name':'drug_name'})
meds['drug_exposure_start_datetime'] = meds['drug_exposure_start_datetime'].astype(str)
meds[['issue_date', 'extra']] = meds.drug_exposure_start_datetime.str.split(' ', expand = True)
meds

# Select columns we need and sort by date
meds = meds[['ID', 'drug_name', 'issue_date']]
meds = meds.sort_values(by = 'issue_date')
meds

# only keep meds from 1999 or later
meds[['year', 'month', 'day']] = meds.issue_date.str.split('-', expand = True)
meds['year'] = meds['year'].astype(int)
meds = meds[meds['year'] > 1998]
meds

# Select columns we need
meds = meds[['ID', 'drug_name', 'issue_date']]
meds

#Drop duplicated drug names
drugs = meds.drop_duplicates(subset = 'drug_name', keep = 'first')
drugs = drugs[['drug_name']]

#Make everything lower case
drugs['new_name'] = drugs['drug_name'].str.lower()

#Remove all spaces -- necessary because there are sometimes leading or trailing spaces
drugs['new_name2'] = drugs['new_name'].str.replace(' ', '')

print(drugs.dtypes)

# This loads the file we created of keywords to omit
import pandas as pd
s1 = pd.read_csv('Supplementary_Table_1_ Medication_short_name_table - omit_keyword_n100.csv')

s = pd.concat([s1])

#make sure cleaned_med and short_name are both lower case
s = s[~s['short_name'].isna()]
s = s[~s['cleaned_med'].isna()]

#Make everything lower case
s['short_name'] = s['short_name'].str.lower()
s['cleaned_med'] = s['cleaned_med'].str.lower()

#Remove spaces
s['short_name'] = s['short_name'].str.replace(' ', '')
s['cleaned_med'] = s['cleaned_med'].str.replace(' ', '')
s

#Select columns
s = s[['cleaned_med', 'short_name']]
list_drugs = list(s['cleaned_med'])
print(len(list_drugs))
print(list_drugs[0:5])
s

# Create a regular expression pattern to match any of the strings in 'strings_to_check'
pattern = '|'.join(list_drugs)

# Extract the matched string into a new column
matches = drugs['new_name2'].str.extract(f'({pattern})', expand=False)
 
# Print the extracted matches
print(matches)

# Assign the extracted matches to a new column
drugs['omit_keyword'] = matches

# look at df
drugs

# only select drug names that do NOT have an omitted keyword
drugs = drugs[drugs['omit_keyword'].isna()]
drugs

# Load list of all the drugs we tested that were significant
sig = pd.read_csv('Supplementary_Table_2_ cox_baseline_all_results - all_drugs_tested.csv')
sig = sig[sig['found_in_AoU']!='no']
sig = sig[sig['found_in_AoU']!= 'combo']

# This is an inhaler -- should have been removed
sig = sig[sig['PRIOR']!='ipratropium']
list_drugs = list(sig['found_in_AoU'])
print(len(list_drugs))

# Create a regular expression pattern to match any of the strings in 'strings_to_check'
pattern = '|'.join(list_drugs)

# Extract the matched string into a new column
matches = drugs['new_name2'].str.extract(f'({pattern})', expand=False)
 
# Print the extracted matches
print(matches)

# Assign the extracted matches to a new column
drugs['cleaned_med'] = matches

# Check df with cleaned_med names
drugs

# Only keep drugs that matched
match = drugs[~drugs['cleaned_med'].isna()]
match

# Look at med value_counts
match.cleaned_med.value_counts()

# This table we will use to create a cleaned_med column for our original list of records
new = match[['new_name', 'cleaned_med']]
new

# For meds df, make the drug_name lowercase
meds['drug_name'] = meds['drug_name'].str.lower()
meds

# Merge with cleaned_med table
meds2 = meds.merge(new, left_on = 'drug_name', right_on = 'new_name', how = 'left')
meds2

# only keep meds that have a cleaned_med name
meds2 = meds2[~meds2['cleaned_med'].isna()]
meds2

# save file
meds2.to_csv('meds_with_cleaned_name_MAY_05_2025.csv', header = True, index = False)

# Create list of drugs
list_drugs = list(set(list(meds2['cleaned_med'])))
print(len(list_drugs))

# For each drug in the list_drugs, create a separate csv file that includes: 
# 1) the first date an individual was prescribed that medication
# 2) the N for the number of times that medication appears in their EHR to proxy exposure
for drug in list_drugs:
    test = meds2[meds2['cleaned_med'] == drug]      
    test = test.rename(columns = {'issue_date': drug + '_DATE'})
    test = test[['ID', drug + '_DATE', 'drug_name', 'cleaned_med']]
    test2 = pd.DataFrame(test.ID.value_counts()).reset_index()
    test2 = test2.rename(columns = {'count':drug + '_N'})
    test = test.sort_values(by = drug + '_DATE')
    test = test.drop_duplicates(subset = 'ID', keep = 'first')
    test = test.merge(test2, left_on = 'ID', right_on = 'ID', how = 'left')
    test = test[~test[drug + '_DATE'].isna()]
    test.to_csv(f'{drug}_with_date.csv', header = True, index = False)

# Look at one such file
test

# Remove combo drugs for acetaminophen by searching for "/"
drug = 'acetaminophen'
test = pd.read_csv(f'{drug}_with_date.csv')
test['contains_test'] = test['drug_name'].str.contains('/', case=False, na=False, regex = False)
test

# Remove combos
test = test[test['contains_test'] == False]
test

test.to_csv(f'{drug}_with_date.csv', header = True, index = None)

# Remove combo drugs for codeine by searching for "/"
drug = 'codeine'
test = pd.read_csv(f'{drug}_with_date.csv')
test['contains_test'] = test['drug_name'].str.contains('/', case=False, na=False, regex = False)
test


# Remove combos
test = test[test['contains_test'] == False]
test

test.to_csv(f'{drug}_with_date.csv', header = True, index = None)

import os
import subprocess
import numpy as np
import pandas as pd

# This snippet assumes you run setup first

# This code saves your dataframe into a csv file in a "data" folder in Google Bucket

for drug in list_drugs:

    # Replace df with THE NAME OF YOUR DATAFRAME
    my_dataframe = pd.read_csv(f'{drug}_with_date.csv')

    # Replace 'test.csv' with THE NAME of the file you're going to store in the bucket (don't delete the quotation marks)
    destination_filename = f'{drug}_with_date.csv'

    ########################################################################
    ##
    ################# DON'T CHANGE FROM HERE ###############################
    ##
    ########################################################################

    # save dataframe in a csv file in the same workspace as the notebook
    my_dataframe.to_csv(destination_filename, index=False)

    # get the bucket name
    my_bucket = os.getenv('WORKSPACE_BUCKET')

    # copy csv file to the bucket
    args = ["gsutil", "cp", f"./{destination_filename}", f"{my_bucket}/data/drugs/"]
    output = subprocess.run(args, capture_output=True)

    # print output from gsutil
    output.stderr

# This snippet assumes that you run setup first

# This code lists objects in your Google Bucket

# Get the bucket name
my_bucket = os.getenv('WORKSPACE_BUCKET')

# List objects in the bucket
print(subprocess.check_output(f"gsutil ls -lh -r {my_bucket}", shell=True).decode('utf-8'))
