==> All_of_Us_Researcher_Workbench/06_AoU_pull_icd10.ipynb <==
# Pull ICD10 codes

import pandas
import os

# This query represents dataset "icd10_codes" for domain "condition" and was generated for All of Us Controlled Tier Dataset v8
dataset_77125449_condition_sql = """
    SELECT
        c_occurrence.person_id,
        c_occurrence.condition_concept_id,
        c_standard_concept.concept_name as standard_concept_name,
        c_standard_concept.concept_code as standard_concept_code,
        c_standard_concept.vocabulary_id as standard_vocabulary,
        c_occurrence.condition_start_datetime,
        c_occurrence.condition_end_datetime,
        c_occurrence.condition_type_concept_id,
        c_type.concept_name as condition_type_concept_name,
        c_occurrence.stop_reason,
        c_occurrence.visit_occurrence_id,
        visit.concept_name as visit_occurrence_concept_name,
        c_occurrence.condition_source_value,
        c_occurrence.condition_source_concept_id,
        c_source_concept.concept_name as source_concept_name,
        c_source_concept.concept_code as source_concept_code,
        c_source_concept.vocabulary_id as source_vocabulary,
        c_occurrence.condition_status_source_value,
        c_occurrence.condition_status_concept_id,
        c_status.concept_name as condition_status_concept_name 
    FROM
        ( SELECT
            * 
        FROM
            `""" + os.environ["WORKSPACE_CDR"] + """.condition_occurrence` c_occurrence 
        WHERE
            (
                condition_source_concept_id IN (SELECT
                    DISTINCT c.concept_id 
                FROM
                    `""" + os.environ["WORKSPACE_CDR"] + """.cb_criteria` c 
                JOIN
                    (SELECT
                        CAST(cr.id as string) AS id       
                    FROM
                        `""" + os.environ["WORKSPACE_CDR"] + """.cb_criteria` cr       
                    WHERE
                        concept_id IN (1326482, 1326483, 1326492, 1326493, 1326506, 1326588, 1326589, 1326590, 1326591, 1326600, 1326601, 1326602, 1326603, 1326604, 1326605, 1326606, 1326607, 1326608, 1326609, 1553751, 1553752, 1553753, 1553754, 1553766, 1553767, 1553768, 1553769, 1553771, 1553772, 1553773, 1553774, 1553776, 1553777, 1553778, 1553779, 1553781, 1553782, 1553783, 1553784, 1567356, 1567357, 1567394, 1567940, 1567943, 1567944, 1567949, 1567956, 1567958, 1567959, 1567960, 1567964, 1567965, 1567966, 1567969, 1567971, 1567999, 1568062, 1568206, 1568207, 1568211, 1568213, 1568214, 1568217, 1568218, 1568219, 1568221, 1568222, 1568232, 1568233, 1568236, 1568240, 1568242, 1568300, 1568322, 1568325, 1568327, 1568329, 1568330, 1568337, 1568339, 1568344, 1568362, 1568363, 1568364, 1568366, 1568989, 1569124, 1569125, 1569126, 1569127, 1569133, 1569134, 1569135, 1569136, 1569145, 1569170, 1569171, 1569172, 1569173, 1569174, 1569175, 1569176, 1569177, 1569178, 1569179, 1569180, 1569184,
 1569187, 1569192, 1569193, 1569220, 1569221, 1569225, 1569254, 1569348, 1569526, 1569528, 1569558, 1569562, 1569577, 1569619, 1569620, 1569622, 1569630, 1569632, 1569633, 1569650, 1569669, 1569670, 1569671, 1569680, 1569681, 1569772, 1569777, 1569778, 1569783, 1570039, 1570303, 1570328, 1570329, 1570333, 1570612, 1570654, 1571057, 1571059, 1571060, 1571061, 1571062, 1571063, 1571064, 1571077, 1571085, 1571286, 1571473, 1571486, 1571495, 1571496, 1571497, 1571515, 1571516, 1571518, 1595534, 1595535, 1595597, 1595598, 1595600, 1595601, 1595617, 1595618, 1595619, 1595620, 35205413, 35205414, 35205415, 35205416, 35205417, 35205418, 35205419, 35205420, 35205421, 35205422, 35205737, 35205738, 35205739, 35205740, 35205741, 35205820, 35205821, 35205822, 35205823, 35205824, 35205825, 35205826, 35205827, 35206878, 35206879, 35206881, 35206882, 35206927, 35206928, 35206929, 35206930, 35206931, 35206932, 35206933, 35207060, 35207061, 35207062, 35207063, 35207064, 35207065, 35207066, 35207067,
 35207092, 35207093, 35207094, 35207095, 35207096, 35207097, 35207098, 35207099, 35207100, 35207126, 35207127, 35207128, 35207129, 35207130, 35207131, 35207145, 35207146, 35207147, 35207148, 35207149, 35207150, 35207151, 35207152, 35207153, 35207154, 35207155, 35207156, 35207157, 35207158, 35207159, 35207160, 35207161, 35207167, 35207168, 35207169, 35207170, 35207171, 35207172, 35207173, 35207174, 35207175, 35207176, 35207177, 35207178, 35207179, 35207180, 35207181, 35207182, 35207183, 35207184, 35207185, 35207186, 35207187, 35207188, 35207189, 35207190, 35207191, 35207192, 35207193, 35207194, 35207195, 35207196, 35207197, 35207198, 35207199, 35207200, 35207409, 35207410, 35207668, 35207669, 35207670, 35207671, 35207672, 35207675, 35207676, 35207677, 35207678, 35207679, 35207680, 35207681, 35207682, 35207683, 35207684, 35207685, 35207702, 35207703, 35207704, 35207705, 35207706, 35207780, 35207781, 35207782, 35207783, 35207784, 35207785, 35207786, 35207787, 35207788, 35207789, 35207790,
 35207791, 35207792, 35207793, 35207803, 35207804, 35207805, 35207806, 35207807, 35207808, 35207809, 35207810, 35207811, 35207812, 35207813, 35207814, 35207815, 35207816, 35207817, 35207818, 35207819, 35207820, 35207821, 35207822, 35207823, 35207824, 35207825, 35207826, 35207827, 35207828, 35207829, 35207830, 35207831, 35207832, 35207833, 35207834, 35207835, 35207836, 35207837, 35207888, 35207889, 35207891, 35208134, 35208135, 35208136, 35208138, 35208139, 35208140, 35208141, 35208142, 35208143, 35208144, 35208145, 35208188, 35208189, 35208190, 35208191, 35208192, 35208193, 35208194, 35208195, 35208196, 35208197, 35208198, 35208199, 35208201, 35208202, 35208203, 35208204, 35208205, 35208206, 35208207, 35208208, 35208209, 35208238, 35208239, 35208240, 35208241, 35208242, 35208243, 35208244, 35208245, 35208246, 35208288, 35208289, 35208290, 35208291, 35208292, 35208293, 35208330, 35208331, 35208332, 35208333, 35208334, 35208335, 35208337, 35208338, 35208339, 35208340, 35208341, 35208342,
 35208343, 35208344, 35208345, 35208346, 35208347, 35208348, 35208349, 35208350, 35208351, 35208352, 35208353, 35208354, 35208355, 35208356, 35208358, 35208359, 35208360, 35208361, 35208362, 35208363, 35208364, 35208365, 35208366, 35208368, 35208474, 35208475, 35208476, 35208477, 35208478, 35208479, 35208480, 35208481, 35208505, 35208506, 35208507, 35208508, 35208509, 35208510, 35208511, 35208535, 35208536, 35208537, 35208538, 35208539, 35208540, 35208541, 35208542, 35208543, 35208751, 35208752, 35208753, 35208758, 35208759, 35208760, 35208761, 35208762, 35208763, 35208764, 35208765, 35208766, 35208767, 35208768, 35208769, 35208770, 35208771, 35208772, 35208773, 35208774, 35208775, 35208776, 35208777, 35208778, 35208779, 35208780, 35208825, 35208826, 35208827, 35209005, 35209006, 35209007, 35209008, 35209009, 35209010, 35209011, 35209013, 35209038, 35209039, 35209040, 35209112, 35209113, 35209205, 35209206, 35209207, 35209208, 35209210, 35209211, 35209212, 35209213, 35209214, 35209246,
 35209274, 35209275, 35209276, 35209277, 35209278, 35209279, 35209280, 35209301, 35209302, 35209303, 35209304, 35209305, 35209306, 35209307, 35209308, 35209309, 35209310, 35209324, 35209325, 35209326, 35209327, 35209482, 35209483, 35209484, 35209485, 35209486, 35209487, 35209488, 35209489, 37200141, 37200142, 37200143, 37200144, 37200145, 37200146, 37200147, 37200148, 37200149, 37200150, 37200151, 37200152, 37200153, 37200154, 37200155, 37200156, 37200157, 37200158, 37200159, 37200160, 37200161, 37200162, 37200163, 37200164, 37200165, 37200166, 37200167, 37200168, 37200170, 37200171, 37200172, 37200175, 37200176, 37200177, 37200180, 37200181, 37200182, 37200185, 37200186, 37200187, 37200188, 37200189, 37200190, 37200191, 37200192, 37200194, 37200195, 37200196, 37200198, 37200199, 37200200, 37200201, 37200202, 37200203, 37200204, 37200205, 37200206, 37200207, 37200208, 37200209, 37200210, 37200211, 37200212, 37200213, 37200214, 37200215, 37200216, 37200217, 37200218, 37200219, 37200220,
 37200221, 37200222, 37200223, 37200224, 37200225, 37200227, 37200228, 37200229, 37200230, 37200232, 37200233, 37200234, 37200235, 37200237, 37200238, 37200239, 37200240, 37200242, 37200243, 37200244, 37200245, 37200246, 37200247, 37200248, 37200249, 37200251, 37200252, 37200253, 37200254, 37200312, 37200313, 37200319, 37200320, 37200323, 37200324, 37200325, 37200326, 37200327, 37200329, 37200496, 37200497, 37200498, 37200499, 37200500, 37200501, 37200502, 37200503, 37200504, 37200505, 37200506, 37200507, 37200508, 37200509, 37200510, 37200511, 37200512, 37200513, 37200514, 37200515, 37200518, 37200519, 37200520, 37200521, 37200522, 37200523, 37200524, 37200526, 37200527, 37200528, 37200529, 37200530, 37200531, 37200532, 37200536, 37200537, 37200538, 37200539, 37200540, 37200541, 37200542, 37200543, 37200544, 37200545, 37200546, 37200547, 37200548, 37200549, 37200550, 37200551, 37200552, 37200553, 37200554, 37200555, 37200556, 37200557, 37200558, 37200559, 37200560, 37200561, 37200586,
 37200587, 37200588, 37200589, 37200590, 37200591, 37200592, 37200593, 37200594, 37200595, 37200596, 37200597, 37200598, 37200599, 37200600, 37200601, 37200602, 37200603, 37200649, 37200650, 37200651, 37200652, 37200877, 37200878, 37200928, 37200929, 37200930, 37200931, 37402336, 37402337, 37402367, 37402368, 37402369, 37402370, 37402471, 37402472, 37402489, 37402490, 37402491, 37402492, 37402498, 37402503, 37402504, 37402533, 45533017, 45533018, 45533019, 45533020, 45533021, 45533022, 45533023, 45533102, 45533103, 45533104, 45533110, 45533111, 45533135, 45533136, 45533137, 45533138, 45533140, 45533147, 45533148, 45533149, 45533389, 45533390, 45533391, 45533436, 45533439, 45533455, 45533456, 45533457, 45533459, 45533460, 45533462, 45533463, 45533464, 45533466, 45533467, 45533468, 45533472, 45533473, 45533475, 45533477, 45533478, 45533479, 45533480, 45533481, 45533508, 45533509, 45533510, 45533511, 45533512, 45533513, 45533514, 45533515, 45533516, 45533517, 45533518, 45533519, 45533520,
 45533567, 45533586, 45533602, 45533605, 45533606, 45533616, 45533640, 45533701, 45533702, 45533703, 45533735, 45533736, 45533737, 45533745, 45533746, 45533747, 45533748, 45533749, 45533840, 45533841, 45533939, 45533940, 45533941, 45533942, 45533943, 45533944, 45533945, 45533946, 45533947, 45533948, 45533949, 45533950, 45533952, 45533954, 45533957, 45533958, 45533959, 45534130, 45537750, 45537958, 45537960, 45537961, 45537962, 45538061, 45538062, 45538063, 45538064, 45538067, 45538068, 45538069, 45538071, 45538105, 45538107, 45538108, 45538109, 45538110, 45538111, 45538112, 45538114, 45538115, 45538121, 45538122, 45538123, 45538334, 45538335, 45538336, 45538337, 45538373, 45538374, 45538375, 45538376, 45538377, 45538388, 45538389, 45538390, 45538392, 45538394, 45538395, 45538396, 45538397, 45538403, 45538405, 45538408, 45538409, 45538410, 45538411, 45538412, 45538413, 45538416, 45538418, 45538419, 45538420, 45538451, 45538452, 45538453, 45538454, 45538455, 45538456, 45538457, 45538508,
 45538510, 45538521, 45538531, 45538532, 45538533, 45538534, 45538535, 45538545, 45538635, 45538639, 45538676, 45538677, 45538681, 45538682, 45538683, 45538685, 45538686, 45538687, 45538688, 45538689, 45538866, 45538867, 45538868, 45538869, 45538870, 45538872, 45538880, 45539039, 45539041, 45539042, 45539047, 45539076, 45542548, 45542736, 45542737, 45542738, 45542748, 45542780, 45542781, 45542841, 45542842, 45542843, 45542844, 45542847, 45542852, 45542881, 45542882, 45542884, 45542885, 45542887, 45542893, 45543118, 45543119, 45543167, 45543168, 45543182, 45543185, 45543187, 45543188, 45543189, 45543190, 45543195, 45543196, 45543197, 45543198, 45543199, 45543201, 45543202, 45543203, 45543204, 45543207, 45543208, 45543229, 45543230, 45543231, 45543232, 45543233, 45543234, 45543235, 45543236, 45543237, 45543238, 45543289, 45543317, 45543318, 45543335, 45543338, 45543339, 45543341, 45543437, 45543438, 45543439, 45543440, 45543442, 45543443, 45543475, 45543484, 45543485, 45543486, 45543487,
 45543577, 45543689, 45543690, 45543691, 45543692, 45543693, 45543694, 45543697, 45543698, 45543699, 45543700, 45543701, 45543705, 45543799, 45543800, 45543860, 45543861, 45543863, 45547621, 45547622, 45547623, 45547624, 45547625, 45547626, 45547627, 45547657, 45547696, 45547703, 45547704, 45547706, 45547707, 45547708, 45547711, 45547731, 45547732, 45547734, 45547735, 45547743, 45547744, 45548010, 45548011, 45548012, 45548013, 45548022, 45548026, 45548027, 45548028, 45548029, 45548030, 45548034, 45548036, 45548039, 45548040, 45548043, 45548045, 45548046, 45548047, 45548048, 45548049, 45548050, 45548051, 45548053, 45548055, 45548080, 45548081, 45548082, 45548083, 45548084, 45548085, 45548136, 45548149, 45548160, 45548161, 45548170, 45548196, 45548197, 45548268, 45548269, 45548270, 45548271, 45548313, 45548314, 45548315, 45548316, 45548321, 45548322, 45548323, 45548324, 45548325, 45548326, 45548418, 45548419, 45548503, 45548504, 45548507, 45548509, 45548510, 45548513, 45548516, 45548517,
 45548523, 45548525, 45548603, 45548653, 45548657, 45552206, 45552207, 45552219, 45552379, 45552381, 45552382, 45552383, 45552385, 45552386, 45552417, 45552477, 45552479, 45552480, 45552489, 45552490, 45552524, 45552527, 45552528, 45552530, 45552537, 45552538, 45552539, 45552741, 45552796, 45552799, 45552801, 45552802, 45552803, 45552804, 45552805, 45552806, 45552807, 45552809, 45552813, 45552817, 45552818, 45552819, 45552820, 45552822, 45552851, 45552852, 45552853, 45552854, 45552908, 45552923, 45552924, 45552930, 45552931, 45552932, 45552933, 45552935, 45552941, 45552946, 45552947, 45552948, 45553051, 45553052, 45553053, 45553089, 45553090, 45553095, 45553097, 45553098, 45553099, 45553184, 45553271, 45553273, 45553274, 45553277, 45553278, 45553280, 45553281, 45553283, 45553285, 45553289, 45553290, 45553390, 45553391, 45553440, 45553444, 45556927, 45557110, 45557111, 45557112, 45557113, 45557201, 45557205, 45557208, 45557228, 45557229, 45557230, 45557233, 45557234, 45557236, 45557237,
 45557238, 45557240, 45557494, 45557495, 45557536, 45557538, 45557539, 45557548, 45557549, 45557551, 45557552, 45557553, 45557554, 45557555, 45557556, 45557557, 45557558, 45557559, 45557561, 45557562, 45557563, 45557564, 45557565, 45557566, 45557567, 45557597, 45557598, 45557599, 45557600, 45557647, 45557657, 45557658, 45557664, 45557672, 45557675, 45557762, 45557763, 45557764, 45557807, 45557808, 45557809, 45557992, 45557993, 45557994, 45557998, 45558001, 45558004, 45558005, 45558148, 45561738, 45561947, 45561948, 45561949, 45561985, 45561986, 45562038, 45562045, 45562046, 45562054, 45562055, 45562087, 45562088, 45562093, 45562095, 45562096, 45562309, 45562340, 45562343, 45562344, 45562353, 45562355, 45562357, 45562359, 45562362, 45562363, 45562364, 45562365, 45562367, 45562370, 45562371, 45562373, 45562377, 45562378, 45562379, 45562380, 45562381, 45562382, 45562383, 45562384, 45562404, 45562405, 45562406, 45562407, 45562408, 45562409, 45562410, 45562411, 45562412, 45562492, 45562493,
 45562494, 45562507, 45562593, 45562594, 45562597, 45562598, 45562599, 45562600, 45562638, 45562639, 45562640, 45562641, 45562642, 45562649, 45562650, 45562651, 45562709, 45562837, 45562838, 45562839, 45562840, 45562841, 45562842, 45562843, 45562844, 45562845, 45562846, 45562847, 45562848, 45562849, 45562851, 45562852, 45562854, 45562855, 45562856, 45562857, 45562858, 45563013, 45563016, 45566545, 45566547, 45566729, 45566731, 45566839, 45566840, 45566841, 45566843, 45566844, 45566846, 45566847, 45566848, 45566849, 45566870, 45566871, 45566872, 45566874, 45566875, 45566876, 45566884, 45566885, 45566886, 45566887, 45567125, 45567128, 45567167, 45567168, 45567180, 45567181, 45567185, 45567187, 45567188, 45567189, 45567191, 45567192, 45567193, 45567195, 45567196, 45567197, 45567198, 45567199, 45567200, 45567201, 45567202, 45567203, 45567204, 45567205, 45567206, 45567234, 45567235, 45567236, 45567237, 45567238, 45567239, 45567240, 45567241, 45567242, 45567243, 45567244, 45567308, 45567309,
 45567310, 45567311, 45567324, 45567420, 45567422, 45567423, 45567424, 45567425, 45567426, 45567459, 45567460, 45567461, 45567462, 45567469, 45567471, 45567472, 45567473, 45567474, 45567545, 45567561, 45567659, 45567660, 45567661, 45567662, 45567665, 45567666, 45567667, 45567668, 45567672, 45567678, 45567780, 45567842, 45571654, 45571753, 45571756, 45571761, 45571762, 45571766, 45571798, 45571799, 45571800, 45571801, 45571802, 45571803, 45571804, 45571805, 45571810, 45571811, 45571812, 45572041, 45572079, 45572080, 45572094, 45572097, 45572098, 45572099, 45572102, 45572104, 45572105, 45572109, 45572110, 45572111, 45572113, 45572114, 45572115, 45572138, 45572139, 45572140, 45572141, 45572142, 45572143, 45572144, 45572146, 45572212, 45572213, 45572214, 45572219, 45572226, 45572340, 45572341, 45572342, 45572344, 45572345, 45572346, 45572368, 45572372, 45572373, 45572374, 45572375, 45572552, 45572557, 45572559, 45572560, 45572561, 45572563, 45572565, 45572567, 45572569, 45572703, 45572704,
 45576437, 45576438, 45576439, 45576440, 45576443, 45576480, 45576538, 45576539, 45576540, 45576541, 45576544, 45576545, 45576546, 45576547, 45576572, 45576573, 45576574, 45576590, 45576591, 45576592, 45576593, 45576820, 45576821, 45576822, 45576865, 45576866, 45576868, 45576876, 45576878, 45576884, 45576885, 45576886, 45576887, 45576888, 45576889, 45576890, 45576891, 45576892, 45576893, 45576894, 45576895, 45576896, 45576897, 45576898, 45576899, 45576900, 45576901, 45576922, 45576923, 45576924, 45576925, 45576926, 45576927, 45576928, 45576929, 45576971, 45576972, 45576989, 45576990, 45576991, 45576999, 45577104, 45577105, 45577106, 45577107, 45577108, 45577109, 45577112, 45577114, 45577115, 45577116, 45577117, 45577118, 45577119, 45577153, 45577154, 45577163, 45577164, 45577165, 45577168, 45577358, 45577359, 45577362, 45577363, 45577366, 45577367, 45577369, 45577371, 45577372, 45577373, 45577467, 45577517, 45577545, 45581141, 45581142, 45581161, 45581162, 45581349, 45581350, 45581352,
 45581353, 45581354, 45581355, 45581459, 45581461, 45581463, 45581487, 45581489, 45581490, 45581491, 45581492, 45581497, 45581498, 45581499, 45581500, 45581727, 45581728, 45581729, 45581730, 45581731, 45581732, 45581766, 45581781, 45581782, 45581783, 45581784, 45581785, 45581787, 45581789, 45581790, 45581794, 45581795, 45581797, 45581819, 45581820, 45581821, 45581822, 45581823, 45581824, 45581825, 45581912, 45582017, 45582049, 45582050, 45582051, 45582052, 45582057, 45582059, 45582061, 45582062, 45582063, 45582064, 45582126, 45582150, 45582247, 45582248, 45582249, 45582250, 45582251, 45582254, 45582257, 45582258, 45582260, 45585944, 45585945, 45586138, 45586139, 45586140, 45586173, 45586174, 45586241, 45586242, 45586243, 45586283, 45586286, 45586287, 45586289, 45586290, 45586291, 45586292, 45586297, 45586298, 45586299, 45586531, 45586533, 45586572, 45586573, 45586574, 45586575, 45586587, 45586588, 45586591, 45586593, 45586594, 45586595, 45586596, 45586597, 45586598, 45586601, 45586602,
 45586603, 45586604, 45586605, 45586610, 45586611, 45586612, 45586634, 45586635, 45586636, 45586637, 45586693, 45586695, 45586717, 45586718, 45586719, 45586721, 45586722, 45586723, 45586730, 45586731, 45586768, 45586840, 45586841, 45586842, 45586843, 45586844, 45586845, 45586885, 45586886, 45586889, 45586890, 45586891, 45586892, 45586893, 45586894, 45586969, 45587064, 45587065, 45587066, 45587067, 45587068, 45587069, 45587070, 45587071, 45587072, 45587073, 45587074, 45587075, 45587076, 45587077, 45587079, 45587081, 45587082, 45587083, 45587088, 45587090, 45587092, 45587192, 45590840, 45590852, 45591026, 45591027, 45591029, 45591030, 45591031, 45591134, 45591135, 45591136, 45591171, 45591173, 45591175, 45591178, 45591181, 45591185, 45591186, 45591414, 45591415, 45591456, 45591458, 45591459, 45591460, 45591461, 45591468, 45591469, 45591471, 45591474, 45591476, 45591477, 45591478, 45591482, 45591489, 45591491, 45591492, 45591520, 45591521, 45591522, 45591523, 45591524, 45591525, 45591571,
 45591579, 45591585, 45591594, 45591701, 45591702, 45591703, 45591705, 45591737, 45591738, 45591739, 45591740, 45591743, 45591744, 45591745, 45591746, 45591747, 45591820, 45591934, 45591935, 45591936, 45591939, 45591940, 45591941, 45591942, 45591944, 45591945, 45591949, 45591953, 45591954, 45591955, 45591956, 45591958, 45592066, 45592126, 45592133, 45592134, 45592135, 45592160, 45595618, 45595793, 45595794, 45595795, 45595797, 45595798, 45595799, 45595903, 45595910, 45595911, 45595936, 45595937, 45595938, 45595940, 45595942, 45595947, 45596156, 45596157, 45596158, 45596197, 45596198, 45596199, 45596212, 45596213, 45596214, 45596215, 45596217, 45596221, 45596222, 45596223, 45596224, 45596225, 45596226, 45596250, 45596251, 45596252, 45596253, 45596254, 45596255, 45596256, 45596257, 45596258, 45596259, 45596294, 45596328, 45596330, 45596331, 45596332, 45596333, 45596373, 45596442, 45596443, 45596444, 45596446, 45596481, 45596482, 45596483, 45596484, 45596485, 45596495, 45596497, 45596498,
 45596687, 45596688, 45596690, 45596691, 45596698, 45596700, 45596800, 45596801, 45596864, 45596889, 45600636, 45600637, 45600638, 45600639, 45600640, 45600641, 45600642, 45600742, 45600743, 45600744, 45600746, 45600747, 45600749, 45600753, 45600778, 45600779, 45600781, 45600782, 45600783, 45600785, 45600786, 45600793, 45600991, 45600992, 45600993, 45600994, 45601024, 45601026, 45601027, 45601028, 45601038, 45601041, 45601042, 45601043, 45601044, 45601045, 45601046, 45601047, 45601048, 45601049, 45601050, 45601051, 45601052, 45601056, 45601058, 45601059, 45601060, 45601061, 45601062, 45601063, 45601064, 45601088, 45601089, 45601090, 45601091, 45601092, 45601093, 45601094, 45601095, 45601140, 45601154, 45601155, 45601177, 45601178, 45601182, 45601187, 45601222, 45601284, 45601285, 45601286, 45601287, 45601288, 45601289, 45601344, 45601349, 45601354, 45601356, 45601433, 45601434, 45601454, 45601545, 45601546, 45601547, 45601548, 45601549, 45601550, 45601551, 45601552, 45601553, 45601555,
 45601557, 45601558, 45601559, 45601562, 45601563, 45601565, 45601567, 45601668, 45601669, 45601727, 45601728, 45601761, 45605231, 45605397, 45605398, 45605401, 45605402, 45605403, 45605404, 45605405, 45605416, 45605500, 45605502, 45605537, 45605538, 45605539, 45605540, 45605541, 45605542, 45605543, 45605545, 45605546, 45605547, 45605556, 45605558, 45605559, 45605560, 45605739, 45605779, 45605784, 45605785, 45605786, 45605787, 45605788, 45605800, 45605801, 45605802, 45605805, 45605806, 45605807, 45605809, 45605814, 45605816, 45605817, 45605818, 45605819, 45605821, 45605822, 45605823, 45605825, 45605827, 45605848, 45605849, 45605850, 45605851, 45605852, 45605853, 45605854, 45605855, 45605856, 45605857, 45605858, 45605911, 45605923, 45605925, 45605932, 45605939, 45605940, 45605952, 45605957, 45606053, 45606054, 45606055, 45606056, 45606057, 45606059, 45606061, 45606062, 45606063, 45606064, 45606115, 45606116, 45606117, 45606118, 45606119, 45606120, 45606126, 45606131, 45606132, 45606133,
 45606214, 45606323, 45606324, 45606327, 45606331, 45606334, 45606336, 45606342, 45606487, 45606489, 45606490, 725299, 725359, 725360, 725361, 725362, 725363, 725364, 725365, 725366, 725367, 725368, 725369, 725372, 725373, 725378, 725379, 725380, 725401, 725402, 725403, 725404, 725407, 725409, 725410, 725414, 725440, 725441, 725442, 766349, 766365, 766366, 766367, 766369, 766371, 766372, 766373, 766374, 766375, 766376, 766378, 766379, 766380, 766402, 766404, 766406, 766408, 766409, 766410, 766411, 9833)       
                        AND full_text LIKE '%_rank1]%'      ) a 
                        ON (c.path LIKE CONCAT('%.', a.id, '.%') 
                        OR c.path LIKE CONCAT('%.', a.id) 
                        OR c.path LIKE CONCAT(a.id, '.%') 
                        OR c.path = a.id) 
                WHERE
                    is_standard = 0 
                    AND is_selectable = 1)
            )) c_occurrence 
    LEFT JOIN
        `""" + os.environ["WORKSPACE_CDR"] + """.concept` c_standard_concept 
            ON c_occurrence.condition_concept_id = c_standard_concept.concept_id 
    LEFT JOIN
        `""" + os.environ["WORKSPACE_CDR"] + """.concept` c_type 
            ON c_occurrence.condition_type_concept_id = c_type.concept_id 
    LEFT JOIN
        `""" + os.environ["WORKSPACE_CDR"] + """.visit_occurrence` v 
            ON c_occurrence.visit_occurrence_id = v.visit_occurrence_id 
    LEFT JOIN
        `""" + os.environ["WORKSPACE_CDR"] + """.concept` visit 
            ON v.visit_concept_id = visit.concept_id 
    LEFT JOIN
        `""" + os.environ["WORKSPACE_CDR"] + """.concept` c_source_concept 
            ON c_occurrence.condition_source_concept_id = c_source_concept.concept_id 
    LEFT JOIN
        `""" + os.environ["WORKSPACE_CDR"] + """.concept` c_status 
            ON c_occurrence.condition_status_concept_id = c_status.concept_id"""

dataset_77125449_condition_df = pandas.read_gbq(
    dataset_77125449_condition_sql,
    dialect="standard",
    use_bqstorage_api=("BIGQUERY_STORAGE_API_ENABLED" in os.environ),
    progress_bar_type="tqdm_notebook")

dataset_77125449_condition_df.head(5)

# Rename df
df = dataset_77125449_condition_df
df

# select columns and split datetime
df = df[['person_id', 'standard_concept_name', 'condition_start_datetime', 'condition_source_value']]
df = df.sort_values(by = 'condition_start_datetime')
df[['condition_start_datetime', 'extra']] = df.condition_start_datetime.astype(str).str.split(' ', expand = True)

# drop extra column
df = df.drop(columns = ['extra'])
df

# Select codes from 1999 or later
df[['year', 'month', 'day']] = df.condition_start_datetime.str.split('-', expand = True)
df['year'] = df['year'].astype(int)
df = df[df['year'] >= 1999]
df

# Select columns needed
df = df[['person_id', 'standard_concept_name', 'condition_start_datetime',
       'condition_source_value']]
df

# Save for backup
df.to_csv('cleaner_list.csv', header = True, index = False)

# Reload list if needed
import pandas as pd
df = pd.read_csv('cleaner_list.csv')
df

# Split ICD10 code to get groups
df[['icd10_group', 'extra']] = df.condition_source_value.str.split('.', expand = True)
df

# Create a list of ICD10 codes
code_list = list(set(list(df['icd10_group'])))
print(len(code_list))
sorted_list = sorted(code_list)
print(sorted_list)

# Save each ICD10 code as its own small csv file
for group in code_list:
    test = df[df['icd10_group']==group]
    print(group, len(test))
    test.to_csv(f'{group}_with_date.csv', header = True, index = None)

# Other eating disorders

# This pulls binge eating disorder, which was coded differently that the other eating disorders,
# even though it also includes an ICD10 code
import pandas
import os

# This query represents dataset "icd10_codes" for domain "observation" and was generated for All of Us Controlled Tier Dataset v8
dataset_77125449_observation_sql = """
    SELECT
        observation.person_id,
        observation.observation_concept_id,
        o_standard_concept.concept_name as standard_concept_name,
        o_standard_concept.concept_code as standard_concept_code,
        o_standard_concept.vocabulary_id as standard_vocabulary,
        observation.observation_datetime,
        observation.observation_type_concept_id,
        o_type.concept_name as observation_type_concept_name,
        observation.value_as_number,
        observation.value_as_string,
        observation.value_as_concept_id,
        o_value.concept_name as value_as_concept_name,
        observation.qualifier_concept_id,
        o_qualifier.concept_name as qualifier_concept_name,
        observation.unit_concept_id,
        o_unit.concept_name as unit_concept_name,
        observation.visit_occurrence_id,
        o_visit.concept_name as visit_occurrence_concept_name,
        observation.observation_source_value,
        observation.observation_source_concept_id,
        o_source_concept.concept_name as source_concept_name,
        o_source_concept.concept_code as source_concept_code,
        o_source_concept.vocabulary_id as source_vocabulary,
        observation.unit_source_value,
        observation.qualifier_source_value,
        observation.value_source_concept_id,
        observation.value_source_value,
        observation.questionnaire_response_id 
    FROM
        ( SELECT
            * 
        FROM
            `""" + os.environ["WORKSPACE_CDR"] + """.observation` observation 
        WHERE
            (
                observation_source_concept_id IN (35207829, 37200328)
            )) observation 
    LEFT JOIN
        `""" + os.environ["WORKSPACE_CDR"] + """.concept` o_standard_concept 
            ON observation.observation_concept_id = o_standard_concept.concept_id 
    LEFT JOIN
        `""" + os.environ["WORKSPACE_CDR"] + """.concept` o_type 
            ON observation.observation_type_concept_id = o_type.concept_id 
    LEFT JOIN
        `""" + os.environ["WORKSPACE_CDR"] + """.concept` o_value 
            ON observation.value_as_concept_id = o_value.concept_id 
    LEFT JOIN
        `""" + os.environ["WORKSPACE_CDR"] + """.concept` o_qualifier 
            ON observation.qualifier_concept_id = o_qualifier.concept_id 
    LEFT JOIN
        `""" + os.environ["WORKSPACE_CDR"] + """.concept` o_unit 
            ON observation.unit_concept_id = o_unit.concept_id 
    LEFT JOIN
        `""" + os.environ["WORKSPACE_CDR"] + """.visit_occurrence` v 
            ON observation.visit_occurrence_id = v.visit_occurrence_id 
    LEFT JOIN
        `""" + os.environ["WORKSPACE_CDR"] + """.concept` o_visit 
            ON v.visit_concept_id = o_visit.concept_id 
    LEFT JOIN
        `""" + os.environ["WORKSPACE_CDR"] + """.concept` o_source_concept 
            ON observation.observation_source_concept_id = o_source_concept.concept_id"""

dataset_77125449_observation_df = pandas.read_gbq(
    dataset_77125449_observation_sql,
    dialect="standard",
    use_bqstorage_api=("BIGQUERY_STORAGE_API_ENABLED" in os.environ),
    progress_bar_type="tqdm_notebook")

dataset_77125449_observation_df.head(5)

# Rename df
df1 = dataset_77125449_observation_df
df1

# select columns we need
df1 = df1[['person_id', 'standard_concept_name', 'observation_datetime', 'observation_source_value']]
df1 = df1.sort_values(by = 'observation_datetime')

#split datetime
df1[['observation_datetime', 'extra']] = df1.observation_datetime.astype(str).str.split(' ', expand = True)
df1

# Drop extra column
df1 = df1.drop(columns = ['extra'])
df1

# only keep codes from 1999 or later
df1[['year', 'month', 'day']] = df1.observation_datetime.str.split('-', expand = True)
df1['year'] = df1['year'].astype(int)
df1 = df1[df1['year'] >= 1999]
df1

# remame/reorder columns keeping the original ICD10 code
df1 = df1[['person_id', 'standard_concept_name', 'observation_datetime',
       'observation_source_value']]
df1 = df1.rename(columns = {'observation_datetime':'condition_start_datetime', 'observation_source_value':'condition_source_value'})
df1

# Split ICD10 code into larger group
df1[['icd10_group', 'extra']] = df1.condition_source_value.str.split('.', expand = True)
df1

# Check value counts
df1.icd10_group.value_counts()

df1.standard_concept_name.value_counts()

# Reload previously created F50 file to combine data
import pandas as pd
test = pd.read_csv('F50_with_date.csv')
test

# Concat with previously created F50 file
df = pd.concat([test, df1])
df

df.to_csv('F50_with_date.csv', header = True, index = False)

import os
import subprocess
import numpy as np
import pandas as pd

# This snippet assumes you run setup first

# This code saves your dataframe into a csv file in a "data" folder in Google Bucket

for group in code_list:

    # Replace df with THE NAME OF YOUR DATAFRAME
    my_dataframe = pd.read_csv(f'{group}_with_date.csv')

    # Replace 'test.csv' with THE NAME of the file you're going to store in the bucket (don't delete the quotation marks)
    destination_filename = f'{group}_with_date.csv'

    ########################################################################
    ##
    ################# DON'T CHANGE FROM HERE ###############################
    ##
    ########################################################################

    # save dataframe in a csv file in the same workspace as the notebook
    my_dataframe.to_csv(destination_filename, index=False)

    # get the bucket name
    my_bucket = os.getenv('WORKSPACE_BUCKET')

    # copy csv file to the bucket
    args = ["gsutil", "cp", f"./{destination_filename}", f"{my_bucket}/data/icd10/"]
    output = subprocess.run(args, capture_output=True)

    # print output from gsutil
    output.stderr
